+++ Module2/src/product_selling/service/iCustomerService.java
@@ -1,7 +1,5 @@
 package product_selling.service;
 
-import product_selling.entity.Product;
-
 public interface iCustomerService {
-    public void addToCart(int productId, int quantity);
+    public void addToCart(int productId, double quantity);
 }
